refactor(iroh-net): move `Conn` toplevel
and avoid leaking the `conn::Inner` type
