+++ iroh-net/src/magicsock/derp_actor.rs
@@ -17,7 +17,7 @@ use crate::{
     key::{self, node::PUBLIC_KEY_LENGTH},
 };
 
-use super::conn::{ActorMessage, Inner};
+use super::{ActorMessage, Inner};
 
 /// How long a non-home DERP connection needs to be idle (last written to) before we close it.
 const DERP_INACTIVE_CLEANUP_TIME: Duration = Duration::from_secs(60);
